Jupyter Notebook
The stakeholder of a game corporation believes top reviewers have a specific set of skills that would be very useful in benefiting our platform. They’d like to start an inventive program, where to give the best reviewers early access to games and certain perks for giving such great reviews. Initially, the stakeholder suggests to use 'votes helpful' as the metric to define top reviewers. Reviewers with high numbers of votes are more likely to be valuable for the platform. We also agree on 1000 would be an ideal threshold for the size of top reviewer cohort. The goal of this project is to find the top reviewer cohort, so we can understand the population a bit better, and use a more data-driven approach for targeting.
With the goal in mind, I use SQL to query the database and joined two related tables, author_info table and review_info, together. The complied dataset is about a million rows with 22 columns. After adjusting some column name and data type, I bring the dataset local and use this Jupyter Notebook to create an interactive report within the tight deadline.
In the analysis, I found over 90% of reviewers only published one review, including the reviewer with highest vote_helpful counts. They are one hit wonders, although their reviews go viral and are absolutely worth for investigation, they may not be the right targets for this incentive program, because one hit wonder is less likely to be replicable and what we want is to encourage constant behaviours for user engagement. I create a Pareto chart to prove the idea and bring 'review counts per reviewer' as the second metric for evaluation. 
In the notebook, I also create sliders which enable the report users to do what if analysis. They can adjust constrains of 'vote counts' and 'review counts per reviewer' based on their business knowledge and find the cohort size based on different combinations. What's more, the report dynamically visualizes further characteristics on the given cohort. From reviewer standpoint, it shows the distributions of number of games owned by reviewers and playtime per game reviewed. From review standpoint, the report illustrates character length of  an average review and average comments per review has.
I learnt a lot from this project!